requirements.txt
awscli==1.18.93
boto3==1.14.16
click==7.1.2
PyYAML==5.3.1
